requirements/base.txt
# Django
#-------------------------------------------------------------------------------
Django==2.1.4
django-environ==0.4.5

wagtail==2.4
python-ldap==3.2.0
